refactor : Sign in 과정 data class 네이밍, 위치 변경 #PAR-213
