+++ core/model/src/main/java/online/partyrun/partyrunapplication/core/model/signin/GoogleIdToken.kt
@@ -1,8 +1,5 @@
 package online.partyrun.partyrunapplication.core.model.signin
 
-import com.google.gson.annotations.SerializedName
-
 data class GoogleIdToken(
-    @SerializedName("idToken")
     val idToken: String? = ""
 )
