+++ core/model/src/main/java/online/partyrun/partyrunapplication/core/model/signin/SignInToken.kt
@@ -1,0 +1,6 @@
+package online.partyrun.partyrunapplication.core.model.signin
+
+data class SignInToken(
+    val accessToken: String? = "",
+    val refreshToken: String? = ""
+)
