--- core/model/src/main/java/online/partyrun/partyrunapplication/core/model/signin/SignInTokenResult.kt
@@ -1,11 +1,0 @@
-package online.partyrun.partyrunapplication.core.model.signin
-
-import com.google.gson.annotations.SerializedName
-
-data class SignInTokenResult(
-    /* TODO : Name 변경 요 */
-    @SerializedName("accessToken")
-    val accessToken: String,
-    @SerializedName("refreshToken")
-    val refreshToken: String = ""
-)
